backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
supabase==1.2.0
requests==2.31.0
pandas==2.2.1 